角色设置、角色对应的菜单

+++ VocabEase_common/src/main/java/com/javastudy/vocabease_common/service/impl/RoleServiceImpl.java
@@ -1,16 +1,28 @@
 package com.javastudy.vocabease_common.service.impl;
 
+import com.javastudy.vocabease_common.entity.enums.MenuCheckTypeEnum;
 import com.javastudy.vocabease_common.entity.enums.PageSize;
+import com.javastudy.vocabease_common.entity.enums.ResponseCodeEnum;
+import com.javastudy.vocabease_common.entity.po.Account;
 import com.javastudy.vocabease_common.entity.po.Role;
+import com.javastudy.vocabease_common.entity.po.Role2menu;
+import com.javastudy.vocabease_common.entity.query.AccountQuery;
+import com.javastudy.vocabease_common.entity.query.Role2menuQuery;
 import com.javastudy.vocabease_common.entity.query.RoleQuery;
 import com.javastudy.vocabease_common.entity.query.SimplePage;
 import com.javastudy.vocabease_common.entity.vo.PaginationResultVO;
+import com.javastudy.vocabease_common.exception.BusinessException;
+import com.javastudy.vocabease_common.mappers.AccountMapper;
+import com.javastudy.vocabease_common.mappers.Role2menuMapper;
 import com.javastudy.vocabease_common.mappers.RoleMapper;
 import com.javastudy.vocabease_common.service.RoleService;
 import com.javastudy.vocabease_common.utils.StringTools;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 
@@ -22,6 +34,12 @@ public class RoleServiceImpl implements RoleService {
 
 	@Resource
 	private RoleMapper<Role, RoleQuery> roleMapper;
+
+	@Resource
+	private Role2menuMapper<Role2menu, Role2menuQuery> role2menuMapper;
+
+	@Resource
+	private AccountMapper<Account, AccountQuery> accountMapper;
 
 	/**
 	 * 根据条件查询列表
@@ -107,7 +125,10 @@ public class RoleServiceImpl implements RoleService {
 	 */
 	@Override
 	public Role getRoleByRoleId(Integer roleId) {
-		return this.roleMapper.selectByRoleId(roleId);
+		Role role = this.roleMapper.selectByRoleId(roleId);
+		List<Integer> selectMenuIds = this.role2menuMapper.selectMenuIdsByRoleIds(new String[]{String.valueOf(roleId)});
+		role.setMenuIds(selectMenuIds);
+		return role;
 	}
 
 	/**
@@ -122,7 +143,79 @@ public class RoleServiceImpl implements RoleService {
 	 * 根据RoleId删除
 	 */
 	@Override
+	@Transactional(rollbackFor = Exception.class)
 	public Integer deleteRoleByRoleId(Integer roleId) {
-		return this.roleMapper.deleteByRoleId(roleId);
+		AccountQuery accountQuery = new AccountQuery();
+		accountQuery.setRoles(String.valueOf(roleId));
+		Integer count = this.accountMapper.selectCount(accountQuery);
+		if (count > 0)
+			throw new BusinessException("该角色仍被使用，无法进行删除");
+		count = this.roleMapper.deleteByRoleId(roleId);
+		Role2menuQuery role2menuQuery = new Role2menuQuery();
+		role2menuQuery.setRoleId(roleId);
+		this.role2menuMapper.deleteByParam(role2menuQuery);
+		return count;
 	}
+
+	@Override
+	@Transactional(rollbackFor = Exception.class)
+	public void savaRole(Role role, String menuIds, String halfMenuIds) {
+		Integer roleId = role.getRoleId();
+		Boolean addMenu = false;
+		//新增
+		if (roleId == null) {
+			Date date = new Date();
+			role.setCreateTime(date);
+			role.setLastUpdateTime(date);
+			this.roleMapper.insert(role);
+			addMenu = true;
+			roleId = role.getRoleId();
+		}
+		//修改
+		else {
+			role.setCreateTime(null);
+			this.roleMapper.updateByRoleId(role, roleId);
+		}
+		RoleQuery roleQuery = new RoleQuery();
+		roleQuery.setRoleName(role.getRoleName());
+		Integer count = this.findCountByParam(roleQuery);
+		if (count > 1)
+			throw new BusinessException("该角色已存在，请重新设置");
+		if (addMenu) {
+			this.saveRole2Menu(roleId, menuIds, halfMenuIds);
+		}
+	}
+
+	public void saveRole2Menu(Integer roleId, String menuIds, String halfMenuIds) {
+		if (roleId == null || menuIds == null)
+			throw new BusinessException(ResponseCodeEnum.CODE_600);
+		Role2menuQuery role2menuQuery = new Role2menuQuery();
+		role2menuQuery.setRoleId(roleId);
+		this.role2menuMapper.deleteByParam(role2menuQuery);
+		String[] menuIdsArray = menuIds.split(",");
+		String[] halfMenuIdsArray = halfMenuIds.split(",");
+		List<Role2menu> role2menuList = getRole2MenuList(roleId, menuIdsArray, halfMenuIdsArray);
+		if (!role2menuList.isEmpty())
+			this.role2menuMapper.insertBatch(role2menuList);
+	}
+
+	private static List<Role2menu> getRole2MenuList(Integer roleId, String[] menuIdsArray, String[] halfMenuIdsArray) {
+		List<Role2menu> role2MenuList = new ArrayList<>();
+		for (String menuId : menuIdsArray) {
+			Role2menu role2menu = new Role2menu();
+			role2menu.setMenuId(Integer.parseInt(menuId));
+			role2menu.setRoleId(roleId);
+			role2menu.setCheckType(MenuCheckTypeEnum.ALL.getCheckTypeCode());
+			role2MenuList.add(role2menu);
+		}
+		for (String menuId : halfMenuIdsArray) {
+			Role2menu role2menu = new Role2menu();
+			role2menu.setMenuId(Integer.parseInt(menuId));
+			role2menu.setRoleId(roleId);
+			role2menu.setCheckType(MenuCheckTypeEnum.HALF.getCheckTypeCode());
+			role2MenuList.add(role2menu);
+		}
+		return role2MenuList;
+	}
+
 }
